requirements.txt
# 🏦 Financial Trading ETL Pipeline - Production Dependencies
# ================================================================
# Professional-grade Python dependencies for financial data processing
# Last updated: October 2025 | Compatible with Python 3.9+

# ================================
# 🚀 CORE ETL FRAMEWORK
# ================================
apache-airflow==2.8.0              # Workflow orchestration
pyspark==3.5.0                     # Big data processing
pandas==2.1.4                      # Data manipulation
numpy==1.25.2                      # Numerical computing
pyarrow==14.0.1                    # Columnar data format

# ================================
# ☁️ AWS CLOUD & DATA LAKE
# ================================
boto3==1.34.0                      # AWS SDK for S3, EMR, etc
botocore==1.34.0                   # AWS core functionality
s3fs==2023.12.2                    # S3 filesystem interface
apache-airflow-providers-amazon==8.17.0  # AWS Airflow integration

# ================================
# 🏗️ DATA WAREHOUSE & CACHE
# ================================
snowflake-connector-python==3.6.0  # Snowflake data warehouse
redis==5.0.1                       # Caching & message broker

# ================================
# 📊 FINANCIAL DATA SOURCES
# ================================
alpha-vantage==2.3.1               # Stock market data API
requests==2.31.0                   # HTTP client for APIs
yfinance==0.2.18                   # Yahoo Finance backup
websocket-client==1.7.0            # Real-time data streams

# ================================
# 💹 TECHNICAL ANALYSIS
# ================================
ta-lib==0.4.28                     # Technical indicators (RSI, MACD, etc)
scipy==1.11.4                      # Statistical functions
scikit-learn==1.3.2                # Machine learning models

# ================================
# 🗄️ DATABASE & STORAGE
# ================================
sqlalchemy==2.0.25                 # Database ORM
psycopg2-binary==2.9.9             # PostgreSQL driver
snowflake-sqlalchemy==1.5.1        # Snowflake integration

# ================================
# ✅ DATA QUALITY & VALIDATION
# ================================
great-expectations==0.18.8         # Data validation framework
pydantic==2.5.2                    # Schema validation
jsonschema==4.20.0                 # JSON validation

# ================================
# 📈 MONITORING & OBSERVABILITY
# ================================
prometheus-client==0.19.0          # Metrics collection
python-json-logger==2.0.7          # Structured logging
structlog==23.2.0                  # Advanced logging

# ================================
# ⚙️ CONFIGURATION & SECURITY
# ================================
python-dotenv==1.0.0               # Environment variables
pyyaml==6.0.1                      # YAML configuration
cryptography==41.0.8               # Security & encryption

# ================================
# 🔧 UTILITIES & HELPERS
# ================================
click==8.1.7                       # CLI framework
tqdm==4.66.1                       # Progress bars
python-dateutil==2.8.2             # Date parsing
pytz==2023.3.post1                 # Timezone handling

# ================================
# 🧪 TESTING & DEVELOPMENT (Optional)
# ================================
pytest==7.4.4                      # Testing framework
pytest-cov==4.1.0                  # Coverage reporting
black==23.12.1                     # Code formatting
flake8==7.0.0                      # Code linting

# ================================
# 📋 INSTALLATION NOTES
# ================================
# Install TA-Lib dependencies first:
# - Windows: Download TA-Lib from https://www.lfd.uci.edu/~gohlke/pythonlibs/
# - Linux/Mac: sudo apt-get install ta-lib (or brew install ta-lib)
#
# For production deployment:
# pip install -r requirements.txt --no-deps --force-reinstall
#
# For development with Jupyter:
# pip install jupyterlab matplotlib seaborn plotly